+++ include/cloudbreak.bash
@@ -22,7 +22,7 @@ cloudbreak-conf-tags() {
     env-import DOCKER_TAG_CONSUL 0.5
     env-import DOCKER_TAG_REGISTRATOR v5
     env-import DOCKER_TAG_POSTGRES 9.4.1
-    env-import DOCKER_TAG_CLOUDBREAK 1.0.2
+    env-import DOCKER_TAG_CLOUDBREAK 1.0.3
     env-import DOCKER_TAG_CBDB 1.0.0
     env-import DOCKER_TAG_PERISCOPE 0.5.6
     env-import DOCKER_TAG_PCDB 0.5.6