+++ include/circle.bash
@@ -30,6 +30,10 @@ cci-latest-org() {
     : ${user:?}
 
     local latest=$(circle "project/$user/$project/tree/$branch?filter=completed&limit=1" |jq .[0].build_num)
+    if ! [ "$latest" -gt 0 ] 2>/dev/null ; then 
+        error "No artifact found for branch: '$branch' at https://circleci.com/gh/sequenceiq/cloudbreak-deployer"
+        exit 1
+    fi
     debug latest build: $latest
 
     circle  "project/$user/$project/$latest/artifacts" | jq .[].url -r | grep -i $(uname)