+++ include/deployer.bash
@@ -1,6 +1,6 @@
 debug() {
   if [[ "$DEBUG" ]]; then
-      echo "[DEBUG] $*" | gray 1>&2
+      echo -e "[DEBUG] $*" | gray 1>&2
   fi
 }
 
@@ -219,6 +219,7 @@ main() {
     cbd-find-root
     deps-init
 	color-init
+    deps-require sed
     load-profile "$@"
 
     circle-init