+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ### Fixed
 
+- Authentication error with `cloudbreak-shell` and `cloudbreak-shell-quiet` is fixed
+
 ### Added
 
 ### Removed