+++ docs/changelog.md
@@ -14,6 +14,7 @@
 - sequenceiq/cloudbreak image updated to 1.0.3
 
 - debug() function made multiline capable. Use \n in messages
+- refactor bridge ip discovery to run helper docker command only once
 
 ## [v1.0.2] - 2015-08-25