+++ docs/changelog.md
@@ -2,11 +2,22 @@
 
 ### Fixed
 
+- Authentication error with `cloudbreak-shell` and `cloudbreak-shell-quiet` is fixed
+- Command `cbd update <branch>` checks for artifact
+
 ### Added
+
+- binary version of gnu-sed 4.2.2 is now included, to solve lot of osx/busybox issues
+- consul recursor test are added
 
 ### Removed
 
 ### Changed
+- sequenceiq/cloudbreak image updated to 1.0.3
+
+- debug() function made multiline capable. Use \n in messages
+- refactor bridge ip discovery to run helper docker command only once
+- consul recursor handling refactored to be more robust
 
 ## [v1.0.2] - 2015-08-25