+++ docs/changelog.md
@@ -4,16 +4,21 @@
 
 - Authentication error with `cloudbreak-shell` and `cloudbreak-shell-quiet` is fixed
 - Command `cbd update <branch>` checks for artifact
-- Command `cbd start` will execute the migration by default. In case of `cbd start --no-migration` command or if SKIP_DB_MIGRATION_ON_START envvar set to true in Profile, the migration will be skipped
 
 ### Added
 
 - binary version of gnu-sed 4.2.2 is now included, to solve lot of osx/busybox issues
 - consul recursor test are added
+- Command `cbd start` will execute the migration by default. In case of `cbd start --no-migration` command or if SKIP_DB_MIGRATION_ON_START envvar set to true in Profile, the migration will be skipped
+- Using Dns SRV record in our services instead of ambassador
+- Using docker linking system in third party services instead of ambassador
 
 ### Removed
 
+- Full removal of ambassador
+
 ### Changed
+
 - sequenceiq/cloudbreak image updated to 1.0.3
 
 - debug() function made multiline capable. Use \n in messages