+++ docs/changelog.md
@@ -2,13 +2,8 @@
 
 ### Fixed
 
-- Authentication error with `cloudbreak-shell` and `cloudbreak-shell-quiet` is fixed
-- Command `cbd update <branch>` checks for artifact
-
 ### Added
 
-- binary version of gnu-sed 4.2.2 is now included, to solve lot of osx/busybox issues
-- consul recursor test are added
 - Command `cbd start` will execute the migration by default. If SKIP_DB_MIGRATION_ON_START envvar set to true in Profile, the migration will be skipped
 - Using Dns SRV record in our services instead of ambassador
 - Using docker linking system in third party services instead of ambassador
@@ -16,6 +11,22 @@
 ### Removed
 
 - Full removal of ambassador
+
+### Changed
+
+## [v1.0.3] - 2015-09-03
+
+### Fixed
+
+- Authentication error with `cloudbreak-shell` and `cloudbreak-shell-quiet` is fixed
+- Command `cbd update <branch>` checks for artifact
+
+### Added
+
+- binary version of gnu-sed 4.2.2 is now included, to solve lot of osx/busybox issues
+- consul recursor test are added
+
+### Removed
 
 ### Changed